async assets laoding for advances settings management section (#62434)
